Split into regression and validation
